requirements.txt
streamlit
supabase
redis
python-dotenv
streamlit-autorefresh
plotly
pandas
pytz
